requirements.txt
anyio==3.6.2
appdirs==1.4.4
attrs==23.1.0
beautifulsoup4==4.12.2
bidict==0.22.1
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.1.0
cssselect==1.2.0
cycler==0.11.0
DateTime==5.1
dateutils==0.6.12
decorator==4.4.2
dnspython==2.3.0
docopt==0.6.2
Flask==2.3.2
Flask-AutoIndex==0.6.6
Flask-Cors==4.0.0
Flask-Silk==0.2
Flask-SocketIO==5.3.4
Flask-Sockets==0.2.1
Flask-SQLAlchemy==3.0.5
flaskapp==1.0.5
fonttools==4.40.0
frozendict==2.3.8
future==0.18.3
gevent==22.10.2
gevent-websocket==0.10.1
google-api-core==2.11.1
google-api-python-client==2.90.0
google-auth==2.20.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.1
greenlet==2.0.2
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
idna==3.4
imageio==2.31.1
imageio-ffmpeg==0.4.8
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
kiwisolver==1.4.4
lxml==4.9.2
MarkupSafe==2.1.2
matplotlib==3.7.1
moviepy==1.0.3
multitasking==0.0.11
nsepy==0.8
numpy==1.24.2
oauthlib==3.2.2
outcome==1.3.0.post0
packaging==23.1
pandas==2.0.0
parsel==1.8.1
pathlib==1.0.1
peewee==3.17.0
pi==0.1.2
Pillow==9.5.0
pipreqs==0.4.13
proglog==0.1.10
protobuf==4.23.3
pyaes==1.6.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pymongo==4.3.3
pyotp==2.8.0
pyparsing==3.1.0
PySocks==1.7.1
python-dateutil==2.8.2
python-engineio==4.4.1
python-slugify==8.0.1
python-socketio==5.8.0
python-telegram-bot==21.1.1
pytz==2023.3
PyYAML==6.0
requests==2.31.0
requests-oauthlib==1.3.1
rfc3986==1.5.0
rsa==4.9
schedule==1.1.0
selectorlib==0.16.0
selenium==4.16.0
six==1.16.0
slugify==0.0.1
smartapi-python==1.3.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
SQLAlchemy==2.0.16
telebot==0.0.5
telegram==0.0.1
Telethon==1.28.5
text-unidecode==1.3
tqdm==4.65.0
trio==0.23.2
trio-websocket==0.11.1
typing_extensions==4.6.3
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
w3lib==2.1.1
waitress==2.1.2
webencodings==0.5.1
websocket-client==1.5.1
websockets==11.0.3
Werkzeug==2.3.4
wsproto==1.2.0
yarg==0.1.9
yfinance==0.2.31
zope.event==4.6
zope.interface==6.0
